requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
google-genai==1.31.0
python-dotenv==1.1.1
google-cloud-storage==2.18.2
pypdf==4.3.1
pyOpenSSL==25.3.0
